Fixed issues

+++ DesktopModules/Vanjaro/UXManager/Library/Resources/Scripts/Blocks/traits.js
@@ -1644,7 +1644,7 @@ export default (editor, config = {}) => {
 
 				if (typeof value == "string" && value != "") {
 
-					inputValue = parseInt(value);
+					inputValue = parseFloat(value);
 
 					if (typeof trait.attributes.units != 'undefined') {
 						$(trait.attributes.units).each(function (index, option) {
@@ -1682,7 +1682,7 @@ export default (editor, config = {}) => {
 				}
 
 				if (property != '' && unit == 'px')
-					inputValue = parseInt($(component.view.el).css(property));
+					inputValue = parseFloat($(component.view.el).css(property));
 
 				trait.view.$el.find('select').val(unit);
